refactor(Profile): extract out the cloning methods to a concern
